fix(ui): 修复管理面板路由冲突和页面卡死问题
根因：Nuxt 同时存在 problems.vue 和 problems/ 目录时，将 problems.vue 当作父路由，[id]/edit.vue 作为子路由。但 problems.vue 没有 <NuxtPage/>，子路由永不渲染。

修复：
- 将 problems/new.vue → problem-new.vue
- 将 problems/[id]/edit.vue → problem-edit/[id].vue
- 更新 problems.vue 中的链接指向新路径
- middleware 添加 5s 超时防止后端不可达时页面卡死
